feat(engine): support inside properties for multi-instance completion condition
