+++ engine/src/main/java/io/camunda/zeebe/engine/state/appliers/ProcessInstanceElementCompletedApplier.java
@@ -10,6 +10,7 @@ package io.camunda.zeebe.engine.state.appliers;
 import io.camunda.zeebe.engine.processing.deployment.model.element.ExecutableCallActivity;
 import io.camunda.zeebe.engine.state.TypedEventApplier;
 import io.camunda.zeebe.engine.state.immutable.ProcessState;
+import io.camunda.zeebe.engine.state.instance.ElementInstance;
 import io.camunda.zeebe.engine.state.mutable.MutableElementInstanceState;
 import io.camunda.zeebe.engine.state.mutable.MutableEventScopeInstanceState;
 import io.camunda.zeebe.engine.state.mutable.MutableVariableState;
@@ -53,6 +54,15 @@ final class ProcessInstanceElementCompletedApplier
 
     eventScopeInstanceState.deleteInstance(key);
     elementInstanceState.removeInstance(key);
+
+    final var flowScopeInstance = elementInstanceState.getInstance(value.getFlowScopeKey());
+
+    if (flowScopeInstance == null) {
+      return;
+    }
+
+    final var flowScopeElementType = flowScopeInstance.getValue().getBpmnElementType();
+    manageMultiInstance(flowScopeInstance, flowScopeElementType);
   }
 
   private boolean isChildProcess(
@@ -76,6 +86,15 @@ final class ProcessInstanceElementCompletedApplier
       final var variables = variableState.getVariablesAsDocument(key);
       eventScopeInstanceState.triggerEvent(
           parentElementInstanceKey, parentElementInstanceKey, elementId, variables);
+    }
+  }
+
+  private void manageMultiInstance(
+      final ElementInstance flowScopeInstance, final BpmnElementType flowScopeElementType) {
+    if (flowScopeElementType == BpmnElementType.MULTI_INSTANCE_BODY) {
+      // update the numberOfCompletedInstances of the multi-instance body
+      flowScopeInstance.incrementNumberOfCompletedElementInstances();
+      elementInstanceState.updateInstance(flowScopeInstance);
     }
   }
 }
