+++ engine/src/main/java/io/camunda/zeebe/engine/state/appliers/ProcessInstanceElementCompletingApplier.java
@@ -8,11 +8,9 @@
 package io.camunda.zeebe.engine.state.appliers;
 
 import io.camunda.zeebe.engine.state.TypedEventApplier;
-import io.camunda.zeebe.engine.state.instance.ElementInstance;
 import io.camunda.zeebe.engine.state.mutable.MutableElementInstanceState;
 import io.camunda.zeebe.protocol.impl.record.value.processinstance.ProcessInstanceRecord;
 import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
-import io.camunda.zeebe.protocol.record.value.BpmnElementType;
 
 /** Applies state changes for `ProcessInstance:Element_Completing` */
 final class ProcessInstanceElementCompletingApplier
@@ -30,23 +28,5 @@ final class ProcessInstanceElementCompletingApplier
     elementInstanceState.updateInstance(
         elementInstanceKey,
         instance -> instance.setState(ProcessInstanceIntent.ELEMENT_COMPLETING));
-
-    final var flowScopeInstance = elementInstanceState.getInstance(value.getFlowScopeKey());
-
-    if (flowScopeInstance == null) {
-      return;
-    }
-
-    final var flowScopeElementType = flowScopeInstance.getValue().getBpmnElementType();
-    manageMultiInstance(flowScopeInstance, flowScopeElementType);
-  }
-
-  private void manageMultiInstance(
-      final ElementInstance flowScopeInstance, final BpmnElementType flowScopeElementType) {
-    if (flowScopeElementType == BpmnElementType.MULTI_INSTANCE_BODY) {
-      // update the numberOfCompletedInstances of the multi-instance body
-      flowScopeInstance.incrementNumberOfCompletedElementInstances();
-      elementInstanceState.updateInstance(flowScopeInstance);
-    }
   }
 }
