+++ engine/src/main/java/io/camunda/zeebe/engine/state/appliers/ProcessInstanceElementTerminatingApplier.java
@@ -8,11 +8,9 @@
 package io.camunda.zeebe.engine.state.appliers;
 
 import io.camunda.zeebe.engine.state.TypedEventApplier;
-import io.camunda.zeebe.engine.state.instance.ElementInstance;
 import io.camunda.zeebe.engine.state.mutable.MutableElementInstanceState;
 import io.camunda.zeebe.protocol.impl.record.value.processinstance.ProcessInstanceRecord;
 import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
-import io.camunda.zeebe.protocol.record.value.BpmnElementType;
 
 /** Applies state changes for `ProcessInstance:Element_Terminating` */
 final class ProcessInstanceElementTerminatingApplier
@@ -29,23 +27,5 @@ final class ProcessInstanceElementTerminatingApplier
   public void applyState(final long key, final ProcessInstanceRecord value) {
     elementInstanceState.updateInstance(
         key, instance -> instance.setState(ProcessInstanceIntent.ELEMENT_TERMINATING));
-
-    final var flowScopeInstance = elementInstanceState.getInstance(value.getFlowScopeKey());
-
-    if (flowScopeInstance == null) {
-      return;
-    }
-
-    final var flowScopeElementType = flowScopeInstance.getValue().getBpmnElementType();
-    manageMultiInstance(flowScopeInstance, flowScopeElementType);
-  }
-
-  private void manageMultiInstance(
-      final ElementInstance flowScopeInstance, final BpmnElementType flowScopeElementType) {
-    if (flowScopeElementType == BpmnElementType.MULTI_INSTANCE_BODY) {
-      // update the numberOfTerminatedInstances of the multi-instance body
-      flowScopeInstance.incrementNumberOfTerminatedElementInstances();
-      elementInstanceState.updateInstance(flowScopeInstance);
-    }
   }
 }
